requirements.txt
geopy=2.2.0
gmplot=1.4.1
Pillow=9.2.0
